Refactor SocialMediaMessage to shared Dapr.Test.Common.Models
