+++ feed-generator/Program.cs
@@ -6,6 +6,7 @@
 namespace FeedGenerator
 {
     using Dapr.Client;
+    using Dapr.Tests.Common.Models;
     using Microsoft.AspNetCore.Hosting;
     using Microsoft.Extensions.Hosting;
     using Prometheus;
