+++ message-analyzer/message-analyzer.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="prometheus-net" Version="3.5.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\common\common.csproj" />
+  </ItemGroup>
+
 </Project>
